% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_functions.R
\name{callApi}
\alias{callApi}
\title{Call API query and return results as a list}
\usage{
callApi(query, failOn404 = TRUE, failonErrors = TRUE, simplifyVector = FALSE)
}
\arguments{
\item{query}{API query URL}

\item{failonErrors}{TRUE/FALSE whether errors (but not 404) should cause a log_fatal and stop()}

\item{simplifyVector}{Parameter of fromJSON function controlling if fromJSON function should return atomic vector instead of list. TRUE = returns atomic vector, FALSE = returns list}

\item{fail_on_404}{TRUE/FALSE whether an 404 API error causes a log_fatal and stop().}
}
\value{
List of results
}
\description{
Function to an API query, get content and convert to a list or atomic vector of results.
}
