% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniprotData.R
\name{callUniprotProteinAPI}
\alias{callUniprotProteinAPI}
\title{Run UniProt API}
\usage{
callUniprotProteinAPI(searchID, species, callType)
}
\arguments{
\item{searchID}{Gene identifier, symbol or primary ID, to search.}

\item{species}{species to search}

\item{callType}{whether to search using the database type or by taxonomy. 'Database', 'Taxonomy' or "Accession"}
}
\value{
Result of the API call as data.frame
}
\description{
Run one of 2 types of UniProt protein APIs for a singular search term. 
\itemize{
\item{1. Database search - Get UniProt entries by using UniProt cross reference(search term's database origin) and ID search term. Takes the searchID along with the species database identifier which relates to the databases Intermine uses. If searchID starts with "^ENSMUSG", the Emsembl database will be used to search.
https://www.ebi.ac.uk/proteins/api/proteins/database:searchID?offset=0&size=-1}
\item{2. Taxonomy search - Get UniProt entries by searching using a exact_gene search term and taxonomy ID. The searchID inputted in to the function will be edited to remove "Dmel_" and "CELE_" from identifier before searching. 
https://www.ebi.ac.uk/proteins/api/proteins?offset=0&size=-1&exact_gene=seachID&taxid=taxid}
}
The function will filter any results to match the taxonomy of the species searched, this is because UniProt can return multiple species for one searchID.
}
