% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_functions.R
\name{checkClass}
\alias{checkClass}
\title{Check parameter is given class}
\usage{
checkClass(parameter, value, classType, functionName)
}
\arguments{
\item{parameter}{The name of the parameter to be checked for use in the error message}

\item{value}{The value to have class checked - the parameter value}

\item{classType}{The class which the function should check for}

\item{functionName}{The name of the function which this function is called from for use in the error message}
}
\description{
Check a given value is expected class and throw fatal error if not
}
