% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_functions.R
\name{checkColumns}
\alias{checkColumns}
\title{Check if columns are missing from data frames}
\usage{
checkColumns(dt, columns, dtParameter, functionName)
}
\arguments{
\item{dt}{A data.frame to check}

\item{columns}{The required columns to be in the data.frame}

\item{dtParameter}{Name of the parameter of the data.frame of the function}

\item{functionName}{The name of the function which this function is called from for use in the error message}
}
\description{
Throw error if columns are missing from input data.frame
}
