% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGeneExists.R
\name{checkGenesExists}
\alias{checkGenesExists}
\title{Check Gene Symbols Exist}
\usage{
checkGenesExists(genes)
}
\arguments{
\item{genes}{character vector of gene symbols to be searched}
}
\value{
data.frame of search gene, column Found (TRUE, FALSE, Synonym) and column Synonym giving the synonym gene symbol if available.
}
\description{
Search HumanMine to determine if gene symbols are found or if synonyms exist,
}
\details{
This function calls HumanMines Gene_description template to search for a list of input genes. The template will search using a lookup so if a gene symbol is a synonym of another gene a result will still be return.
This function will identify if the input genes are found as the official gene symbol, a synonym of another gene (if so return the synonym) or if the gene is not found in form of a data.frame output.
}
