% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrthologues.R
\name{checkOrthologueDataFrame}
\alias{checkOrthologueDataFrame}
\title{Check data is expected orthologue data}
\usage{
checkOrthologueDataFrame(orthologueData, functionName)
}
\arguments{
\item{orthologueData}{Data to be checked for class and columns.}
}
\description{
This function checks data is a data.frame of following columns \itemize{
\item {Gene.primaryIdentifier}
\item {Gene.symbol}
\item {Gene.homologues.homologue.primaryIdentifier}
\item {Gene.homologues.homologue.symbol}
\item {Gene.homologues.homologue.organism.shortName}
\item {Gene.homologues.type}
}
}
