% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_functions.R
\name{checkSpecies}
\alias{checkSpecies}
\title{Check passed species}
\usage{
checkSpecies(species, functionName, includeHuman = FALSE)
}
\arguments{
\item{species}{Character vector of passed species}

\item{functionName}{The name of the function which this function is called from for use in the error message}

\item{includeHuman}{Boolean determining whether the function should allow "Human" as an input or not and whether Human should be added to output if species is NULL.}
}
\value{
Vector of species, if NULL is passed to the function all available species are set.
}
\description{
Check species passed to function are within the accepted list of species.
}
