% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutputs.R
\name{convert_to_percent}
\alias{convert_to_percent}
\title{Convert counts to percentages}
\usage{
convert_to_percent(count_summary)
}
\arguments{
\item{count_summary}{data.frame of read in 'count summary' sheet from G2P-SCAN counts output}
}
\value{
data.frame of count_summary with additional species-percent columns
}
\description{
takes in data.frame of count_summary from G2P-SCAN counts file and calculates gene, entity and reaction percentages relative to human for each pathway and species.
}
