% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateReactomeResource.R
\name{createReactomeHierarchyDataTree}
\alias{createReactomeHierarchyDataTree}
\title{Update and write Reactome Pathway Hierarchy data.tree}
\usage{
createReactomeHierarchyDataTree(version, dataLocation = ".")
}
\arguments{
\item{version}{Version number of Reactome used for file name output}

\item{dataLocation}{The directory path of where the 'g2pScanData' folder should be found}
}
\value{
Full file path of newly written data.tree file
}
\description{
Create and save Reactome hierarchy data.tree using latest Reactome version data from 'https://reactome.org/download/current/ReactomePathwaysRelation.txt'. File is download and read by function and transformed in corresponding data.tree hierarchy.
The data.tree has first node 'topNode' which then the most parental Reactome pathways branch from. The data.tree is written as Rdata object file to 'dataLocation/g2pScanData/pathway_hierarchy_datatrees' location. The full file path of this file is returned by the function.
}
