% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeOutputs.R
\name{createSupplementarySheet}
\alias{createSupplementarySheet}
\title{Create supplementary worksheet}
\usage{
createSupplementarySheet(wb, data, sheetName)
}
\arguments{
\item{wb}{The openxlsx workbook object to write to}

\item{data}{Data to write to the workbook}

\item{sheetName}{The name to name the worksheet}
}
\value{
openxlsx wb with new sheet appended onto it
}
\description{
Create workbook object where data behind counts are outputted, including optional genes in pathway, orthologue, proteins and families data.
}
