% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_functions.R
\name{endCluster}
\alias{endCluster}
\title{Stop cluster and clear cluster pkgglobalenv variable to NULL}
\usage{
endCluster()
}
\description{
Stop cluster and clear cluster pkgglobalenv variable to NULL
}
