% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReactomeEntitiesReactions.R
\name{filterEntitiesReactionsPathways}
\alias{filterEntitiesReactionsPathways}
\title{Filter Entities' and Reaction's counts for a list of pathways}
\usage{
filterEntitiesReactionsPathways(dataToFilter, pathways = NULL)
}
\arguments{
\item{dataToFilter}{Data.frame to filter. Rownames should be pathways}

\item{pathways}{List of Reactome pathway identifiers to filter by. When NULL no filter is applied and all pathways are returned}
}
\value{
Filtered entity and reaction matrix
}
\description{
Reads in matrix of all pathway and species (created by updateEntitiesReactionCounts function) and filter for the pathways of interest.
}
