% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReactomeEntitiesReactions.R
\name{filterEntitiesReactionsSpecies}
\alias{filterEntitiesReactionsSpecies}
\title{Filter Entities and Reaction counts for a list of species}
\usage{
filterEntitiesReactionsSpecies(dataToFilter, species = NULL)
}
\arguments{
\item{dataToFilter}{data.frame to filter. Column names must include species with either '_entities' or '_reactions' suffixes.}

\item{species}{List of species to select to run analysis on and filter for, out of: \itemize{
\item {R. norvegicus}
\item {M. musculus}
\item {D. rerio}
\item {C. elegans}
\item {D. melanogaster}
\item {S. cerevisiae}
} 
When NULL no filter is applied and all species are returned}
}
\value{
Filtered entity and reaction matrix
}
\description{
Reads in matrix of all pathway and species (created by updateEntitiesReactionCounts function) and filters for the species of interest.
}
