% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFamilies.R
\name{filterInterProPathwayResults}
\alias{filterInterProPathwayResults}
\title{Filter a single list of InterPro family IDs based on the InterPro family hierarchy}
\usage{
filterInterProPathwayResults(interpro_list, interpro_levels)
}
\arguments{
\item{interpro_list}{A single comma separated string of InterPro IDs to filter. This should a list of family IDs belonging to a single protein.}

\item{interpro_levels}{The InterPro hierarchy data.frame including hierarchy level column and hierarchy group (branch group).}
}
\value{
singe row data frame of a comma delimited list of unfiltered InterPro IDs, and comma delimited list of filtered InterPro ids.
}
\description{
Filters a list of InterPro IDs based on their hierarchy position based on the interpro_levels data.frame.
A list of InterPro IDs are passed to the function as a single comma separated string and only IDs which are a level 2 (first child) for a branch in the hierarchy is kept. Each branch is filtered separately.
If a branch has only a level 1 ID (parent) and no child in the list of InterPro IDs, the level 1 id is kept.
If an InterPro ID is not found in the hierarchy (usually down to different API and hierarchy versions used) then the InterPro ID is kept also.
}
