% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFamilies.R
\name{filterInterProResults}
\alias{filterInterProResults}
\title{Filter InterPro ID results (all proteins) based on hierarchy position for a dataframe of protein families.}
\usage{
filterInterProResults(familyResults, dataLocation = ".")
}
\arguments{
\item{familyResults}{A data.frame of 'protein.accession', 'Gene.homologues.homologue.organism.shortName' and 'family'.}

\item{dataLocation}{The directory path of where the 'g2pScanData' folder should be found}
}
\value{
data.frame of 'protein.accession', 'Gene.homologues.homologue.organism.shortName' and 'filtered_family'.
}
\description{
Each protein family results are filtered to only IDs which are a level 2 (first child) for a branch in the hierarchy is kept. Each branch is filtered separately.
If a branch has only a level 1 ID (parent) and no child in the list of InterPro IDs, the level 1 id is kept. If an InterPro ID is not found in the hierarchy then the InterPro ID is kept also.

Each protein result is filtered separately.
}
\details{
Filtering is done using the InterPro Hierarchy file found in the interpro_hierarchy_with_levels directory of the dataLocation directory.
The file which matches the InterPro API version shall be used or if not available the file will be updated using this version - see ?updateInterproHierarchy.
}
