% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrthologues.R
\name{filterOrthologueLDO}
\alias{filterOrthologueLDO}
\title{Filter orthologues for least divergent orthologues only}
\usage{
filterOrthologueLDO(dataToFilter, species = NULL)
}
\arguments{
\item{dataToFilter}{Orthologue data.frame to filter. Columns should include all of the following \itemize{
\item {Gene.primaryIdentifier}
\item {Gene.symbol}
\item {Gene.homologues.homologue.primaryIdentifier}
\item {Gene.homologues.homologue.symbol}
\item {Gene.homologues.homologue.organism.shortName}
\item {Gene.homologues.type}
}}

\item{species}{List of species to select to run analysis on and filter for, out of: \itemize{
\item {R. norvegicus}
\item {M. musculus}
\item {D. rerio}
\item {C. elegans}
\item {D. melanogaster}
\item {S. cerevisiae}
} 
When NULL no filter is applied and all species are returned}
}
\value{
Filtered orthologue data.frame
}
\description{
Reads in data.frame of orthologues and filters the "Gene.homologues.type" column to only include "least divergent orthologues".
}
