% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniprotData.R
\name{filterUniprotProteins}
\alias{filterUniprotProteins}
\title{Filter UniProt Protein API results}
\usage{
filterUniprotProteins(proteinData)
}
\arguments{
\item{proteinData}{UniProt protein API output for single gene input as a data.frame}
}
\value{
Single UniProt protein accession
}
\description{
Function to take a single gene UniProt protein API result (as a data.frame) and filter results to obtain a singular protein accession based on the following priority choosing order:
\itemize{
\item{1. Reviewed entries (if no reviewed entries are present, the following steps will occur on unreviewed data)}
\item{2. Strongest evidence of existence. High-low: Experimental evidence at protein level, Experimental evidence at transcript level, Protein inferred from homology, Protein predicted, Protein uncertain.}
\item{3. Longest sequence}
\item{4. Version 1 of sequence}
\item{5. The first result of the remaining data}
}
After the filtering down of results, the function will output a singular protein accession number.
}
