% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutputs.R
\name{get_analysis_summary}
\alias{get_analysis_summary}
\title{Return analysis inputs and pathway summary}
\usage{
get_analysis_summary(input_summary)
}
\arguments{
\item{input_summary}{data.frame of read in 'input summary' sheet from G2P-SCAN output}
}
\value{
a list of summarise input and analysis run data
}
\description{
Take data of first sheet of G2P-SCAN excel outputs (input summary) and output the following:
\itemize{
\item{1. n_gene_searched: the number on input genes run in the analysis}
\item{2. n_genes_mapped: the number of input genes mapped to Reactome pathways }
\item{3. n_genes_not_mapped: the number of input genes not mapped to Reactome pathways}
\item{4. n_genes_not_found: number of input genes not found in Reactome}
\item{5. species_run: the selected species run in the analysis}
\item{6. ortho: the selected orthologue filter applied in the analysis}
\item{7. pathway: the selected pathways filters applied in the analysis}
}
}
