% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{getAllCounts}
\alias{getAllCounts}
\title{Get counts for all pathways}
\usage{
getAllCounts(
  inputGenes,
  species,
  pathwayDT,
  genesInPath,
  orthologueMatrix,
  proteinDT,
  proteinMatrix,
  familyDT,
  entitiesReactions
)
}
\arguments{
\item{inputGenes}{The human input genes used to conduct the analysis}

\item{species}{List of species to select to run analysis on and filter for, out of: \itemize{
\item {R. norvegicus}
\item {M. musculus}
\item {D. rerio}
\item {C. elegans}
\item {D. melanogaster}
\item {S. cerevisiae}
} 
When NULL no filter is applied and all species are returned}

\item{pathwayDT}{The output of the getPathways function.}

\item{genesInPath}{The output of the getGenes.}

\item{orthologueMatrix}{The output of getOrthologueMatrix. Recommended to use matrix of primary ids.}

\item{proteinDT}{The output of getProteins}

\item{proteinMatrix}{The output of getProteinMatrix}

\item{familyDT}{The output of getFamilies}

\item{entitiesReactions}{The output of getEntitiesReactionsCounts}
}
\value{
list of dataframes of count types as rows, the species as columns and count as the data. Dataframes in the list are named with the pathway ID.
}
\description{
Loop through all pathways to calculate the following counts:
\itemize{
\item{Pathway Reactome ID - The pathway identifier for the data in the dataframe}
\item{Pathway Reactome Name - The pathway name for the data in the dataframe}
\item{Input Found - The input genes which have been found in that Human pathway or have a orthologue found for in other species}
\item{Input Found Count - The number of input genes which have been found in that Human pathway or have a orthologue found for in other species}
\item{Mapped Human Genes - The number of Human genes which are represented in the orthologues found for each of the species}
\item{Total Gene Count - The total number of unique gene/ orthologues found for the given pathway and species}
\item{Coverage Percentage - The percentage of input genes found in the pathway vs the number of genes in the pathway. Value only given for Human data}
\item{Protein Count - The number of unique proteins found in each species for the given pathway}
\item{Family Count - The number of unique families found in each species for the given pathway}
\item{Proteins Unmapped to families Count - The number of unique proteins without a family ID assigned in each species for the given pathway}
\item{Entity Count - The number of reactome entities associated with the given pathway for each species}
\item{Reaction Count - The number of reactome reactions associated with the given pathway for each species}}
}
