% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{getAllPathwayCountsWb}
\alias{getAllPathwayCountsWb}
\title{Get counts for all pathways}
\usage{
getAllPathwayCountsWb(wb, pathwaysCountsList)
}
\arguments{
\item{wb}{workbook to add worksheet to}

\item{pathwaysCountsList}{List of pathway count data.frames}
}
\value{
wb object with a sheet for each data.frame in pathwayCountsList
}
\description{
Add worksheets to workbook object where the pathway count matrix is given on a separate tab.
}
