% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeOutputs.R
\name{getCountSummaryWb}
\alias{getCountSummaryWb}
\title{Create singular table for all pathways and species in a workbook sheet}
\usage{
getCountSummaryWb(wb, pathwaysCountsList)
}
\arguments{
\item{wb}{The openxlsx workbook object to write to}

\item{pathwaysCountsList}{A list of data.frames of the output of getAllCounts}
}
\value{
openxlsx wb with new sheet appended onto it
}
\description{
Add worksheet to workbook object of all pathways and species counts in one table.
}
