% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{getDataMatrixCount}
\alias{getDataMatrixCount}
\title{Count unique terms in matrix columns}
\usage{
getDataMatrixCount(genesToCount, dataToCount)
}
\arguments{
\item{genesToCount}{A vector of human genes which should be used to filter matrix prior count}

\item{dataToCount}{Matrix of human gene rows and with columns indicating a new entity to counts eg, species. Data can include genes or proteins.}
}
\value{
list of counts
}
\description{
Count the number of unique values per column (species) in a matrix for a given list of Human genes.
}
\details{
Data to be counted will be filtered to the selected list human genes by filtering on row names. Unique values in each columns are then counted for the remaining rows and
the count values are returned as a list in the order of the columns and do not include counts for NA and blank values.

This function has been built so that the data to be counted is a matrix of orthologues or proteins across species (as columns) with the human gene mapped to the orthologues or proteins as row names.
It has been designed so that the genes to filter to data prior counting are a list of genes within a pathway.
This is so that counts for each species' genes/orthologues are returned for a singular pathway.
}
