% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{getDataMatrixUniqueHumanCount}
\alias{getDataMatrixUniqueHumanCount}
\title{Get counts of unique Human genes which have values in each of the species columns and identify if input genes are mapped.}
\usage{
getDataMatrixUniqueHumanCount(inputGenes, genesToCount, dataToCount)
}
\arguments{
\item{inputGenes}{The human input genes used to conduct the analysis}

\item{genesToCount}{A vector of human genes which should be used to filter matrix prior count}

\item{dataToCount}{Matrix of human gene rows and with columns indicating a new entity to counts eg, species. Data can include genes or proteins.}
}
\value{
list of lists of mapped counts, the input genes found and count of input found. Each list is return in the same order as the matrix columns.
}
\description{
Count the number of Human Genes (within the given list of Human genes) which have values associated in each column (species) in a matrix.
}
\details{
Data to be counted will be filtered to the selected list human genes by filtering on row names. The number of human genes with values in each column are counted for the remaining rows and
the count values are returned as a list in the order of the columns. The input genes will be counted per column and the found input genes and count will be returned per column.

This function has been built so that the data to be counted is a matrix of orthologues or proteins across species (as columns) with the human gene mapped to the orthologues or proteins as row names.
It has been designed so that the genes to filter to data prior counting are a list of genes within a pathway.
This is so that counts for each species' mapped Human genes are returned for a singular pathway.
}
