% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReactomeEntitiesReactions.R
\name{getEntitiesReactionsCounts}
\alias{getEntitiesReactionsCounts}
\title{Get Entities and Reaction counts for a list of pathways and species}
\usage{
getEntitiesReactionsCounts(
  pathways = NULL,
  species = NULL,
  onMissedUpdate = "update",
  dataLocation = "."
)
}
\arguments{
\item{pathways}{List of Reactome pathway identifiers to filter by. When NULL no filter is applied and all pathways are returned}

\item{species}{List of species to select to run analysis on and filter for, out of: \itemize{
\item {R. norvegicus}
\item {M. musculus}
\item {D. rerio}
\item {C. elegans}
\item {D. melanogaster}
\item {S. cerevisiae}
} 
When NULL no filter is applied and all species are returned}

\item{dataLocation}{The directory path of where the 'g2pScanData' folder should be found}
}
\value{
Filtered entity and reaction matrix
}
\description{
Reads in matrix of all pathway and species (created by updateEntitiesReactionCounts function) and filter for the pathways and species of interest.
}
