% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFamilies.R
\name{getFamilies}
\alias{getFamilies}
\title{Get families for each protein accession retrieved from getProteins.}
\usage{
getFamilies(
  proteinResult,
  cluster = NULL,
  stopCluster = TRUE,
  dataLocation = "."
)
}
\arguments{
\item{proteinResult}{Data from output of function getProteins}

\item{cluster}{Cluster to run parallel operation on}

\item{stopCluster}{If TRUE 'parallel::stopCluster(cl = cluster)' is run and localVar variable for cluster is set to NULL. If FALSE the cluster is not stopped}

\item{dataLocation}{The directory path of where the 'g2pScanData' folder should be found}
}
\value{
data.frame of 'protein.accession', 'Gene.homologues.homologue.organism.shortName' and final filtered 'family'.
}
\description{
For every protein row passed to this function in the proteinResult, InterPro's API will be called to pull InterPro family IDs for protein accession. 
Steps:
\itemize{
\item{1. Loop through each non-NA row of the input proteinResult and take protein.accession as the searchID and "Gene.homologues.homologue.organism.shortName" as species of the input to the runInterproSearch function}
\item{2. The runInterproSearch function to run the following API request 'https://www.ebi.ac.uk/interpro/api/entry/interpro/protein/uniprot/accession' for single searchID.}
\item{3. Results from the API call are filtered to only include those with an InterPro type of 'Family', excluding superfamilies and domains.}
\item{4. Results from each API call are combine into one large dataframe}
\item{5. Each protein family results are filtered to only IDs which are a level 2 (first child) for a branch in the hierarchy is kept. Each branch is filtered separately.
 If a branch has only a level 1 ID (parent) and no child in the list of InterPro IDs, the level 1 id is kept. If an InterPro ID is not found in the hierarchy then the InterPro id is kept also.
 Each protein result is filtered separately.}}

If cluster is passed to the function the API calls are run in parallel
}
