% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{getFamilyCount}
\alias{getFamilyCount}
\title{Get counts of mapped and unmapped  families for a set of human genes.}
\usage{
getFamilyCount(genesToCount, familyDT, proteinDT, species)
}
\arguments{
\item{genesToCount}{A vector of human genes which should be used to filter matrix prior count}

\item{familyDT}{The dataframe output of getFamilies function. Columns required: 'filtered_family', 'protein.accession' and 'Gene.homologues.homologue.organism.shortName'}

\item{proteinDT}{The dataframe output of the getProteins function. Columns required: 'Gene.symbol', 'protein.accession' and 'Gene.homologues.homologue.organism.shortName'}

\item{species}{List of species to select to run analysis on and filter for, out of: \itemize{
\item {R. norvegicus}
\item {M. musculus}
\item {D. rerio}
\item {C. elegans}
\item {D. melanogaster}
\item {S. cerevisiae}
} 
When NULL no filter is applied and all species are returned}
}
\value{
list of mapped and unmapped counts
}
\description{
Count the number of unique terms (families) per column (species) and count the number of proteins which do not have terms associated with them
}
\details{
Filter a data matrix of Human gene rows and data types (species) as column for a selection of human genes (pathway genes) and count number of unique ids in each of the columns. Values in columns are separated by commas before counting.
Mapped counts are the number of unique id found per column (species) for the genesToCounts. Unmapped counts are the counts of unique proteins without a family ID assigned.
Counts are returned as a list and do not include counts for NA and blank values.

This function has been built so that the data to be counted is a matrix of families across species (as columns) with the human gene mapped to the protein of which the family is derived from are as row names.
It has been designed so that the genes to filter to data prior counting are a list of genes within a pathway.
This is so that counts for each species' families are returned for a singular pathway.
}
