% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFamilies.R
\name{getFamilyMatrix}
\alias{getFamilyMatrix}
\title{Get matrix of proteins}
\usage{
getFamilyMatrix(familyResult, proteinResult, species)
}
\arguments{
\item{familyResult}{Data from output of function getFamilies}

\item{proteinResult}{Data from output of function getProteins}

\item{species}{List of species to select to run analysis on and filter for, out of: \itemize{
\item {R. norvegicus}
\item {M. musculus}
\item {D. rerio}
\item {C. elegans}
\item {D. melanogaster}
\item {S. cerevisiae}
} 
When NULL no filter is applied and all species are returned}
}
\value{
data.frame matrix of human genes as rows and columns as species,including Human, where families are listed as values.
}
\description{
Create data.frame matrix of human genes as rows and columns as species, including human, where families are listed as values.
}
