% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrthologues.R
\name{getFormatedOrthologueString}
\alias{getFormatedOrthologueString}
\title{Format orthologues for one species and one human gene into single string}
\usage{
getFormatedOrthologueString(
  orthologueData,
  humanGene,
  species,
  orthologueOutput
)
}
\arguments{
\item{orthologueData}{The data.frame from the output of function getOrthologueGenes}

\item{humanGene}{Human gene to get orthologue of}

\item{species}{Species to the the orthologue gene for}

\item{orthologueOutput}{Whether orthologue matrix should use the gene HumanMine "primaryIdentifier" or "geneSymbol" or "both"(default) (not case sensitive). When "both" an orthologue will be outputed as 'primaryIdentifier(geneSymbol)'. Orthologues regularly do not have gene symbols associated with them.}
}
\value{
Formated orthologue for single human gene and species
}
\description{
Create a single orthologue string for use in output representation using the following orthologue information: HumanMine primaryidentifier, gene symbol or both.
This function takes the output of the function 'getOrthologues' and select the singular species orthologue for the given human genes and returns a formatted string with select information type.
}
