% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenes.R
\name{getGenes}
\alias{getGenes}
\title{Get genes in pathways}
\usage{
getGenes(pathways)
}
\arguments{
\item{pathways}{Character vector of Reactome pathways identifiers to be searched}
}
\value{
A data.frame of genes symbol and gene HumanMine identifiers with pathway identifier which the genes belong to.
}
\description{
Get a list of all genes found in a list of pathways (using HumanMine via InterMineR package).
This function take a vector of Reactome pathway identifiers and calls HumanMine's template query 'PathwayGenes' which returns all genes which are found in each of the inputted pathways.
}
