% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEntitiesReactions.R
\name{getHumanEntitiesCounts}
\alias{getHumanEntitiesCounts}
\title{Get Human Entities Counts}
\usage{
getHumanEntitiesCounts(pathway)
}
\description{
Run Reactome content service APIs to extract Human pathway entity counts for a single given pathway identifier.
The API will pulls the data of the live version of Reactome and so this function only needs to be run when the HumanMine instance of Reactome is updated to maintain equivalent versions of Reactome across the tool.
The function calls "https://reactome.org/ContentService/data/participants/path_id/referenceEntities" and filters participants returned to only those from the following databases:
\itemize{
\item{UniProt}
\item{ChEBI}
\item{EMBL}
\item{ENSEMBL}
\item{miRBase}}
 Finally the number of participants returned after filtering is set as the Human entity count for the given pathway.
 This filtering occurs to match those counts produced in the Reactome species comparison tool.
 This function is intended to be run for pathways which no count data is found across other species when updating the entities and reaction file used in the tool.
 @param pathway Reactome pathway identifier of interested
 @importFrom logger log_warn
 @include global_functions.R
 @details 
 The function will not stop when an incorrect pathway identifier is used but instead will return NULL. If there is an error other than 404 error in the API then the function will fail and stop.
 
 @return Human entity count for given pathway
}
