% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVersions.R
\name{getHumanMineOrthologueVersion}
\alias{getHumanMineOrthologueVersion}
\title{Get version number of Panther currently used in HumanMine}
\usage{
getHumanMineOrthologueVersion()
}
\value{
Version number
}
\description{
Get version number of Panther currently used in HumanMine
}
