% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVersions.R
\name{getHumanMineReactomeVersion}
\alias{getHumanMineReactomeVersion}
\title{Get version number of the Reactome database currently used in HumanMine}
\usage{
getHumanMineReactomeVersion()
}
\value{
Version number
}
\description{
Get version number of the Reactome database currently used in HumanMine
}
