% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVersions.R
\name{getHumanMineVersion}
\alias{getHumanMineVersion}
\title{Get version number of HumanMine}
\usage{
getHumanMineVersion()
}
\value{
Version number
}
\description{
Get version number of HumanMine
}
