% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEntitiesReactions.R
\name{getHumanReactionsCounts}
\alias{getHumanReactionsCounts}
\title{Get Human Reactions Counts}
\usage{
getHumanReactionsCounts(pathway)
}
\arguments{
\item{pathway}{Reactome pathway identifier of interested}
}
\value{
Human reaction count for given pathway
}
\description{
Run Reactome content service APIs to extract Human pathway reaction counts for a single given pathway identifier.
The API will pull the data of the live version of Reactome and so this function only needs to be run when the HumanMine instance of Reactome is updated to maintain equivalent versions of Reactome across the tool.
The function calls "https://reactome.org/ContentService/data/pathway/path_id/containedEvents" and filters events returned to only those which have 'className' = Reaction.
This filtering occurs to match those counts produced in the Reactome species comparison tool. Finally the number of events returned after filtering is set as the Human reaction count for the given pathway.
This function is intended to be run for pathways which no count data is found across other species when updating the entities and reaction file used in the tool.
}
\details{
The function will not stop when an incorrect pathway identifier is used but instead will return NULL. If there is an error other than 404 error in the API then the function will fail and stop.
}
