% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEntitiesReactions.R
\name{getHumanSpeciesEntitiesReactionsCounts}
\alias{getHumanSpeciesEntitiesReactionsCounts}
\title{Get A Single Species vs Human Entity and Reaction Counts}
\usage{
getHumanSpeciesEntitiesReactionsCounts(speciesName, speciesCode)
}
\arguments{
\item{speciesName}{Species name to be labeled on data.frame results column}

\item{speciesCode}{Species Reactome identifier code to be used in the API call}
}
\value{
Data.frame of pathway identifier, Human and species entities and reaction counts.
}
\description{
Run Reactome analysis service API to extract Human and a given species' pathway entity and reaction counts.
The API will pull the data of the live version of Reactome and so this function only needs to be run when HumanMine instance of Reactome is updated to maintain equivalent versions of Reactome across the tool.
The function calls "https://reactome.org/AnalysisService/species/homoSapiens/speciesCode?sortBy=ENTITIES_PVALUE&order=ASC&resource=TOTAL&pValue=1" and takes counts of Human and species entities and reactions from the results.
}
