% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInterproHierarchy.R
\name{getInterproLevels}
\alias{getInterproLevels}
\title{Get InterPro family branches and hierarchy levels of InterPro families}
\usage{
getInterproLevels(dt)
}
\arguments{
\item{dt}{A data.frame of the InterPro's 'ParentChildTreeText.txt' (from InterPro ftp site). Quotes should be removed from data in this data.frame. InterPro IDs should be given in column V1 and order of these ids along with "--" indicating hierarchical level must remain.}
}
\value{
data.frame of family identifier, family name, hierarchy level and hierarchy branch number
}
\description{
Build data.frame representation of the InterPro Hierarchy using 'ParentChildTreeText.txt' (from InterPro ftp site) data.frame passed to function. Each family will be given as a row in the data.frame with InterPro ID, name, hierarchy level and family branch/group.
The order of the hierarchy is maintained in the data.frame. A family branch number represents one branch of the family tree and every family with the same family branch number belong to the same most parental family. The level indicates which generation the family is of within the family tree of that branch.
Level 1 means the family is the most parental family. Level 2 means the family is a child of the level 1 family (1st generation child). Level 3 means that the family is the child of a level 2 family (2nd generation child). Levels are defined as number of parents of an individual family plus 1. The number of parents of a family is determined by the number of "--" prefixing the family identifier.
}
