% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVersions.R
\name{getInterProVersions}
\alias{getInterProVersions}
\title{Get version number of InterPro used by the InterPro API and sub databses used by InterPro}
\usage{
getInterProVersions()
}
\value{
List of all InterPro versions, including InterPro itself, the protein instances and member databases with name and version and description.
}
\description{
Get version number of InterPro used by the InterPro API and sub databses used by InterPro
}
