% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrthologues.R
\name{getOrthologueGenes}
\alias{getOrthologueGenes}
\title{Get all orthologues}
\usage{
getOrthologueGenes(genes, species = NULL, orthologueFilter = "LDO")
}
\arguments{
\item{genes}{List of Human gene symbol of which orthologues should be searched}

\item{species}{List of species to select to run analysis on and filter for, out of: \itemize{
\item {R. norvegicus}
\item {M. musculus}
\item {D. rerio}
\item {C. elegans}
\item {D. melanogaster}
\item {S. cerevisiae}
} 
When NULL no filter is applied and all species are returned}

\item{orthologueFilter}{"All" (include all orthologues) or "LDO" (filter to least divergent orthologues)}
}
\value{
Data.frame of Human genes and species orthologues. Columns :\itemize{
\item {Gene.primaryIdentifier}
\item {Gene.symbol}
\item {Gene.homologues.homologue.primaryIdentifier}
\item {Gene.homologues.homologue.symbol}
\item {Gene.homologues.homologue.organism.shortName}
\item {Gene.homologues.type}
}
}
\description{
This function get all orthologues for a list of genes. Species to get orthologues for can be chosen from \itemize{
\item{R. norvegicus}
\item{M. musculus}
\item{D. rerio}
\item{C. elegans}
\item{D. melanogaster}
\item{S. cerevisiae}}
Orthologues are pulled from Panther via HumanMine using HumanMine query template "Gene_Orth". Orthologues which are included can be chosen from the following: \itemize{
\item {All Orthologues - where all orthologues are selected}
\item {Least Divergent Orthologues (LDO) - where only the most nearly 'equivalent' gene in another organism is selected (in case of gene duplications for instance)}
}
}
