% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrthologues.R
\name{getOrthologueMatrix}
\alias{getOrthologueMatrix}
\title{Get matrix of orthologues}
\usage{
getOrthologueMatrix(
  orthologueData,
  species,
  allHumanGenes = NULL,
  orthologueOutput = "both"
)
}
\arguments{
\item{orthologueData}{Data from output of function getOrthologueGenes}

\item{species}{Species to the the orthologue gene for}

\item{allHumanGenes}{Character vector of all human genes symbols to be included in matrix, even if there is no orthologue result in orthologueData. If NULL only genes found in orthologueData will be used.}

\item{orthologueOutput}{Whether orthologue matrix should use the gene HumanMine "primaryIdentifier" or "geneSymbol" or "both"(default) (not case sensitive). When "both" an orthologue will be outputed as 'primaryIdentifier(geneSymbol)'. Orthologues regularly do not have gene symbols associated with them.}
}
\value{
data.frame matrix of human genes as rows and columns as species where orthologues are listed as values.
}
\description{
Create data.frame matrix of human genes as rows and columns as species where orthologues are listed as values.
}
