% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathways.R
\name{getPathwayLevels}
\alias{getPathwayLevels}
\title{Map pathways to Reactome pathway hierarchy}
\usage{
getPathwayLevels(pathways, onMissedUpdate = "update", dataLocation = ".")
}
\arguments{
\item{pathways}{A dataframe including  Reactome pathway "Pathway Identifiers" of which hierarchy levels should returned}

\item{onMissedUpdate}{Character string to indicate what to do when Reactome file cannot be updated to appropriate version.
Either "update" to use the Reactome API most up to date version. "use latest" to use the most up to date version of the file saved in the dataLocation appropriate repository already or given an integer to state what version of Reactome file to use - this must be already present in the dataLocation appropriate repository.}

\item{dataLocation}{The directory path of where the 'g2pScanData' folder should be found}
}
\value{
Pathways dataframe with an appended "Pathway Level" column with 'parental', 'intermediate' or 'terminal' labels for each pathway
}
\description{
Map Reactome pathways to the Reactome hierarchy to get a hierarchy level, 'parental', 'intermediate' or 'terminal', for each pathways passed to the function.
}
\details{
Pathways are mapped to a data.tree of all Reactome pathways using saved hierarchy data.tree Rdata file of Reactome hierarchy structure of all pathways. The data.tree is subset using the pathway identifiers in the pathwayData parameter and the hierarchy of the subset of pathways is maintained.
Each pathway is then labelled with parental, intermediate and terminal levels based on their position in the subset hierarchy. Parental level pathways are defined as those which have children in the hierarchy of selected pathways but do not have any parent themselves.
Terminal level pathways are defined as those which have parent pathways, but no children and intermediate pathways have both parent and children pathways. Hierarchy level is determined based on the filter hierarchy of pathways.

The Rdata file can be found in the pathway_hierarchy_datatrees directory of the dataLocation directory. The file with the matching version to HumanMine version of Reactome is used if exists and if not use or creation of the file is based on theMissedUpdate parameter and updateReactomeHierarchy
}
