% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathways.R
\name{getPathways}
\alias{getPathways}
\title{Get Pathways}
\usage{
getPathways(
  inputGenes,
  pathwayLevels = c("parental", "intermediate", "terminal"),
  onMissedUpdate = "update",
  dataLocation = "."
)
}
\arguments{
\item{inputGenes}{Character vector of gene symbols to be searched}

\item{pathwayLevels}{Character vector indicating what pathways levels should be returned, 'parental','intermediate' and/or 'terminal'. The default value = c('parental','intermediate','terminal') to include all pathways. Any combination of pathway levels can be used.
\itemize{
\item {parental = Pathways which have no parent pathway - the top of Reactome's pathway hierarchy}
\item {intermediate = Pathways which are not parental or terminal and have both parent and children pathways}
\item {terminal = Pathways which have no children pathways - the end of Reactome's pathway hierarchy}
}}

\item{onMissedUpdate}{Character string to indicate what to do when Reactome file cannot be updated to appropriate version.
Either "update" to use the Reactome API most up to date version. "use latest" to use the most up to date version of the file saved in the dataLocation appropriate repository already or given an integer to state what version of Reactome file to use - this must be already present in the dataLocation appropriate repository.}

\item{dataLocation}{The directory path of where the 'g2pScanData' folder should be found}
}
\value{
A data.frame of HumanMine 'Gene_Pathway' query including pathway name and identifier for each gene, link to pathway Reactome site, number of geneInput in pathway and the Reactome hierarchy level (for that subset of pathways).
}
\description{
Get Reactome pathways which gene input are found in (using HumanMine via InterMineR package) and filter human Reactome pathways based on hierarchy level.
}
\details{
This function takes a vector of gene symbols and calls HumanMine's template query 'Gene_Pathway' which returns all pathways genes in the vector belong to.
Pathways results from HumanMine are  filtered for only Human Reactome defined pathways. (Reactome verison used is dependent on verison loaded in HumanMine - see https://www.humanmine.org/humanmine/dataCategories.do).
Pathways are further filtered for only pathways which are the same hierarchy level as set in the pathwayLevels parameter.
The hierarchy level is determined by a subset of the hierarchy data.tree created where the hierarchy levels are as follows:
\itemize{
\item {parental = Pathways which have no parent pathway - the top of Reactome's pathway hierarchy}
\item {intermediate = Pathways which are not parental or terminal and have both parent and children pathways}
\item {terminal = Pathways which have no children pathways - the end of Reactome's pathway hierarchy}
}
The data.tree of all Reactome pathways is read from or saved to a Rdata object found at the pathway_hierarchy_datatrees directory of the dataLocation directory and version of the Reactome hierarchy used is determined by the updateReactomeHierarchy function and onMissedUpdate parameter. The Reactome hierarchy of the same version of Reactome that HumanMine uses will be used when possible.
}
