% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniprotData.R
\name{getProteinMatrix}
\alias{getProteinMatrix}
\title{Get matrix of proteins}
\usage{
getProteinMatrix(proteinResult, orthologueData, species)
}
\arguments{
\item{proteinResult}{Data from output of function getProteins}

\item{orthologueData}{Data from output of function getOrthologueGenes}

\item{species}{List of species to select to run analysis on and filter for, out of: \itemize{
\item {R. norvegicus}
\item {M. musculus}
\item {D. rerio}
\item {C. elegans}
\item {D. melanogaster}
\item {S. cerevisiae}
} 
When NULL no filter is applied and all species are returned}
}
\value{
data.frame matrix of human genes as rows and columns as species,including Human, where proteins are listed as values.
}
\description{
Create data.frame matrix of human genes as rows and columns as species, including human, where protein accessions are listed as values.
}
