% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateReactomeResource.R
\name{getReactomeFileOfVersionPath}
\alias{getReactomeFileOfVersionPath}
\title{Get file path of specified version of Reactome pathway hierarchy tree if available}
\usage{
getReactomeFileOfVersionPath(dataName, folder, version, dataLocation = ".")
}
\arguments{
\item{dataName}{The string of what data is being searched for. Used for logger messages.}

\item{folder}{The folder location to search for files in the package 'data' folder.}

\item{version}{Version number or 'latest' to get specific versions}

\item{dataLocation}{The directory path of where the 'g2pScanData' folder should be found}
}
\value{
File path of Rdata file of Reactome pathway hierarchy if available, NULL if not.
}
\description{
This function looks in the dataLocation/g2pScanData folder and in the directory specified from the 'folder' parameter and searches for data.tree with same version as the 'version' parameter specifies. Versions numbers of the files are determined by the characters after the last underscore in the file name. If a file of matching version is available that file path is returned.
If matching version is not found then NULL is returned.
When parameter version is set to string "latest", then the files are searched and the one with the highest version number has it's file path returned.
}
