% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVersions.R
\name{getUniProtVersion}
\alias{getUniProtVersion}
\title{Get version number of UniProt used by the UniProt API}
\usage{
getUniProtVersion()
}
\value{
UniProt release date
}
\description{
Get version number of UniProt used by the UniProt API
}
