% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVersions.R
\name{getVersionOfReactomeDataUsed}
\alias{getVersionOfReactomeDataUsed}
\title{Get version number of Reactome resources used in analysis}
\usage{
getVersionOfReactomeDataUsed(var)
}
\arguments{
\item{var}{"reactome_datatree" or "reactome_ent_react_file" to pull version of}
}
\value{
Version number
}
\description{
Get version number of Reactome resources used in analysis
}
