% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_functions.R
\name{initiateCluster}
\alias{initiateCluster}
\title{Initiate cluster for parallel compute}
\usage{
initiateCluster(cores)
}
\arguments{
\item{cores}{Number of cores to parallelise}
}
\value{
Built cluster
}
\description{
Set up PSOCK clusters and their environment readyy for use in the G2P-SCAN package
}
