% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutputs.R
\name{plot_entities_reaction_boxplots}
\alias{plot_entities_reaction_boxplots}
\title{Plot entities and Reaction Box Plots}
\usage{
plot_entities_reaction_boxplots(
  count_percent,
  data_name,
  plot_output_dir = "./",
  width = 14,
  height = 6
)
}
\arguments{
\item{count_percent}{data.frame output of convert_to_percent function}

\item{data_name}{string prefixing output png (recommended to match G2P-SCAN file output names)}

\item{plot_output_dir}{output directory to write png files to. If output-dir does not exist, the directory will be created.}

\item{width}{width of output png}

\item{height}{height of output png}
}
\value{
list of entity and reaction ggplot objects
}
\description{
Takes count percentages (produced by convert_to_percent function) and creates a box plot of entity counts across pathways per species and of reaction counts across pathways per species.
}
