% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutputs.R
\name{plot_percent_bubbles}
\alias{plot_percent_bubbles}
\title{Plot bubble plots for gene, entities or reactions}
\usage{
plot_percent_bubbles(
  count_percent,
  count_type = "genes",
  plot_output_dir = "./",
  n_genes_searched,
  data_name,
  join_plots = TRUE,
  n_pathways_per_plot = 22,
  height = 12,
  width = 15
)
}
\arguments{
\item{count_percent}{data.frame output of convert_to_percent function}

\item{count_type}{'genes', 'entities' or 'reactions' indicating what percent type to plot}

\item{plot_output_dir}{output directory to write png files to. If output-dir does not exist, the directory will be created.}

\item{n_genes_searched}{number of genes run in the analysis (output from get_analysis_summary)}

\item{data_name}{string prefixing output png (recommended to match G2P-SCAN file output names)}

\item{join_plots}{TRUE for concatenating bubble and bar plot together, FALSE to produce 2 separate plots}

\item{n_pathways_per_plot}{integer of number of pathways to display in one plot.}

\item{height}{height of output png}

\item{width}{width of output png}
}
\value{
ggplot objects of combined plot or individual plots given in a list
}
\description{
Takes count percentages (produced by convert_to_percent function) and creates a plot displaying percents by bubble size per pathways and species. Bar charts to of pathway sizes can be appended to bubble plot or created separately.
This function can only plot gene, entity or reaction percentages one at a time. 
Due to varying number of pathways outputted from a single G2P-SCAN analysis, number of pathways plotted in one plot is customisable - use 'n_pathways_per_plot' to set this size. If n_pathways_per_plot is lower than total number of pathways, multiple plots will be created and saved, iterating through all pathways.
}
