% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutputs.R
\name{plot_protein_bars}
\alias{plot_protein_bars}
\title{Plot proteins and family bar chart}
\usage{
plot_protein_bars(
  count_percent,
  data_name,
  plot_output_dir = "./",
  n_pathways_per_plot = 25,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{count_percent}{data.frame output of convert_to_percent function}

\item{data_name}{string prefixing output png (recommended to match G2P-SCAN file output names)}

\item{plot_output_dir}{output directory to write png files to. If output-dir does not exist, the directory will be created.}

\item{n_pathways_per_plot}{integer of number of pathways to display in one plot.}

\item{width}{width of output png}

\item{height}{height of output png}
}
\value{
list of ggplot objects produced by the function
}
\description{
Takes count percentages (produced by convert_to_percent function) and creates a plot with a barchart per pathway showing number of proteins, number or families and number of unmapped proteins per species per pathway.
Due to varying number of pathways outputted from a single G2P-SCAN analysis, number of pathways plotted in one plot is customisable - use 'n_pathways_per_plot' to set this size. If n_pathways_per_plot is lower than total number of pathways, multiple plots will be created and saved, iterating through all pathways.
}
