% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutputs.R
\name{run_all_plots}
\alias{run_all_plots}
\title{Wrapper function to run all G2P-SCAN output plots
Given a directory, this function will read each file with suffix "_counts.xlsx" assuming a G2P-SCAN output. For each file the following plots are created:
\itemize{
\item{1. Bubble plots for gene percentage relative to human for each species and pathway}
\item{2. Protein and family bar charts}
\item{3. Entity box plots}
\item{4. Reaction box plots}
}}
\usage{
run_all_plots(g2p_output_dir = "../", plot_output_dir = "./", data_name = NULL)
}
\arguments{
\item{g2p_output_dir}{the directory path for '_counts.xlsx' files to have plots created.}

\item{plot_output_dir}{output directory to write png files to. If output-dir does not exist, the directory will be created.}

\item{data_name}{output file prefix. If left NULL, input file name will be used.}
}
\description{
Wrapper function to run all G2P-SCAN output plots
Given a directory, this function will read each file with suffix "_counts.xlsx" assuming a G2P-SCAN output. For each file the following plots are created:
\itemize{
\item{1. Bubble plots for gene percentage relative to human for each species and pathway}
\item{2. Protein and family bar charts}
\item{3. Entity box plots}
\item{4. Reaction box plots}
}
}
