% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGenes2Pathways.R
\name{runGenes2Pathways}
\alias{runGenes2Pathways}
\title{Run Genes2Pathways pipeline}
\usage{
runGenes2Pathways(
  inputGenes,
  pathwayLevels,
  species = NULL,
  orthologueFilter = "all",
  orthologueOutput = "both",
  cores = (parallel::detectCores() - 1),
  g2pScanData = ".",
  outputDir = ".",
  outputPrefix = "g2pRun",
  pathwaysTabs = TRUE,
  countSummary = TRUE,
  versions = TRUE,
  inputSummary = TRUE,
  reactomeOnMissingUpdate = "update",
  useSynonyms = FALSE
)
}
\arguments{
\item{inputGenes}{Character vector of gene symbols to be searched}

\item{pathwayLevels}{Character vector indicating what pathways levels should be returned, 'parental','intermediate' and/or 'terminal'. The default value = c('parental','intermediate','terminal') to include all pathways. Any combination of pathway levels can be used.
\itemize{
\item {parental = Pathways which have no parent pathway - the top of Reactome's pathway hierarchy}
\item {intermediate = Pathways which are not parental or terminal and have both parent and children pathways}
\item {terminal = Pathways which have no children pathways - the end of Reactome's pathway hierarchy}
}}

\item{species}{List of species to select to run analysis on and filter for, out of: \itemize{
\item {R. norvegicus}
\item {M. musculus}
\item {D. rerio}
\item {C. elegans}
\item {D. melanogaster}
\item {S. cerevisiae}
} 
When NULL no filter is applied and all species are returned}

\item{orthologueFilter}{A character vector of the orthologue filter used for the analysis}

\item{orthologueOutput}{Whether orthologue matrix should use the gene HumanMine "primaryIdentifier" or "geneSymbol" or "both"(default) (not case sensitive). When "both" an orthologue will be outputed as 'primaryIdentifier(geneSymbol)'. Orthologues regularly do not have gene symbols associated with them.}

\item{cores}{The number of cores to run in parallel. If NULL analysis will be run sequentially.}

\item{g2pScanData}{The directory location of the folder "g2pScanData". If this folder does not exist state where this folder of data should be written}

\item{outputDir}{Output directory to write output file to. If directory does not exist, directory will be created at the current working directory}

\item{outputPrefix}{Character vector to prefix all output files}

\item{pathwaysTabs}{Boolean value. TRUE to include a tab per pathway count matrix in counts output}

\item{countSummary}{Boolean value. TRUE to include a table for all counts (all pathways and species) in a singular tab in the counts output}

\item{versions}{Boolean value. TRUE to include version tab}

\item{inputSummary}{Boolean value. TRUE to include a summary of the inputs of the analysis}

\item{reactomeOnMissingUpdate}{Character string to indicate what to do when Reactome files (for pathway hierarchy and entity and reaction counts) cannot be updated to appropriate version.
Either "update" to use the Reactome API most up to date version. "use latest" to use the most up to date version of the file saved in the g2pScanData appropriate repository already or given an integer to state what version of Reactome file to use - this must be already present in the g2pScanData appropriate repository.}

\item{useSynonyms}{Boolean value indicating when an inputted gene string is not identified as predominant identifier for a gene in HumanMine whether the identified main synonym should be used instead. TRUE will use synonyms when dominant gene symbol is found instead.}
}
\description{
The main wrapper function of the G2P-SCAN pipeline. This function runs the G2P-SCAN pipeline from gene input to final counts tables across model species of genes, proteins, families, entities and reactions.
}
\details{
The pipeline is run using the following steps: \enumerate{
\item{Get Pathways \itemize{\item{ 
This is the first step of the pipeline where input genes are mapped to pathways. Using HumanMine via InterMineR to get all Reactome pathways which any of the input genes passed to the function are are found in.
Pathways which have any of the input genes in are filter based on the user pathwayLevel parameter input. The Reactome pathway hierarchy is used to determine hierarchy level of the identified pathways.
The Reactome pathway hierarchy is read in or written to (in case of an update) to the g2pScanData/pathway_hierarchy_datatrees directory and the Reactome version matching HumanMines instance of Reactome will be used when available. }}
}
\item{Get Genes in Pathways \itemize{\item{Taking the identified pathways from step 1, all genes found in those pathways are retrieved using HumanMine's Reactome data via InterMineR queries.}}
}
\item{Get entity and reaction counts for each pathway and each species \itemize{\item{Taking the identified pathways from step 1, the number of entities and reactions for each of the pathways are identified for each chosen model species and Human.
Entities and Reactions are pulled using Reactome's Content and Analysis service APIs and are read in from or written to (in case of an update) the g2pScanData/Reactome_entities_reactions directory and the Reactome version matching HumanMine's instance of Reactome will be used when available.}}
}
\item{Get orthologues of pathway genes for each species \itemize{\item{For all human genes identified in step 2, orthologue genes are pulled using Panther via HumanMine/InterMineR queries.}}
}
\item{Get proteins for identified orthologues \itemize{\item{All species genes identified in step 4, are queried in UniProt to map a protein accession to each gene.
Only a single protein is accepted in the pipeline per gene queried and so, when multiple proteins are returned for one gene query, the following criteria is used:
\enumerate{
\item{Reviewed entries (if no reviewed entries are present, the following steps will occur on unreviewed data)}
\item{Strongest evidence of existence. High-low: Experimental evidence at protein level, Experimental evidence at transcript level, Protein inferred from homology, Protein predicted, Protein uncertain.}
\item{Longest sequence}
\item{Version 1 of sequence}
\item{The first result of the remaining data}
}}}
}
\item{Get Families \itemize{\item{All proteins from step 5 are queried using the InterPro API service to identify associated families to each of the proteins.
Only InterPro entries which are classified as a 'Family' are considered for the pipeline analysis (not 'domain' or 'superfamily'). Families returned for a single proteins are filtered to only include those which are a level 2 (first child) family for a branch in the hierarchy of InterPro families.
The InterPro hierarchy is read in or written to (in case of an update) to the g2pScanData/interpro_hierarchy_with_levels directory.}}
}s
\item{Count and output data \itemize{\item{Pathways - Count the number of pathways which are mapped to the input genes and meet the pathway level criteria}
      \item{Genes - Count the number of unique genes/orthologues in each identified pathway and per species and the number of Human genes which have orthologues mapped to them for each pathway and model species}
      \item{Entities & Reactions - Output entity and reation counts as returned by step 3 individually}
      \item{Proteins - Count the number of unique selected proteins for each identified pathway and species}
      \item{Families - Count the number of unique selected InterPro families or each identified pathway and species and the number of proteins per pathway and species which have no families assigned to them (unassigned)}
      \item{Output - Counts are combined per pathway and all pathways together and excels can be created of counts and background data (for pathways genes, protein and families)}}
}
}

Both the protein and Family steps (steps 5 an 6) can have queries run in parallel if a number of cores is defined. When a numeric value is passed to the cores parameter, PSOCK clusters are created to divide queries and run in parallel. If no cores are defined then queries are run sequentially.
}
