% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFamilies.R
\name{runInterproSearch}
\alias{runInterproSearch}
\title{Get InterPro families for a single protein accession}
\usage{
runInterproSearch(accession, species)
}
\arguments{
\item{accession}{The protein accession to search}

\item{species}{The species to search}
}
\value{
singe row data frame of protein.accession, species and comma delimited list of InterPro ids
}
\description{
Calls the IntePro API with the following API request 'https://www.ebi.ac.uk/interpro/api/entry/interpro/protein/uniprot/accession' for single searchID.
Results from the API call are filtered to only include those with an InterPro type of 'Family', excluding super families and domains. A final comma separated list of family IDs is added to a single row dataframe with the protein accession search term and species.
}
