% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateReactomeResource.R
\name{updateEntitiesReactionsCounts}
\alias{updateEntitiesReactionsCounts}
\title{Get Entities and Reaction count matrix for all Species vs Human}
\usage{
updateEntitiesReactionsCounts(onMissedUpdate = "update", dataLocation = ".")
}
\arguments{
\item{dataLocation}{The directory path of where the 'g2pScanData' folder should be found}

\item{outputDir}{file path to directory where file should be saved}
}
\value{
Entity Reaction matrix
}
\description{
Update and save Reactome entity and reaction counts if the version of Reactome in HumanMine has changed since last data.tree was saved. This function looks in the package's data/Reactome_entities_reactions location and searches for csv with same version of current Reactome in HumanMine (obtained by InterMineR call). If a csv of matching version is available no update is needed and global R.utils option variable reactome_ent_react_file is set as the matching csv full file path.
If matching versioned file is not found, Reactome API is used to get entities and reactions and transformed into the appropriate format. If Reactome version and HumanMine Reactome instance versions do not match, because an update has been missed, there's an option to use the most up to date csv found in the package's data/Reactome_entities_reactions directory or the create a file using the most up to data version from Reactome.
Any new csv created are saved at data/Reactome_entities_reactions of the package directory and this file path is set as the R.utils option variable reactome_ent_react_file of localVars for use throughout the package.
Update of the entities and reaction is performed by running Reactome APIs to extract entity and Reaction counts per pathway for Human and all 6 species to be used in the Genes to pathways tool -
C. elegans, D. rerio, D. melanogaster, M. musculus, R. norvegicus and S. cerevisiae.
}
