% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInterproHierarchy.R
\name{updateInterproHierarchy}
\alias{updateInterproHierarchy}
\title{Update InterPro Hierarchy csv}
\usage{
updateInterproHierarchy(dataLocation = ".")
}
\arguments{
\item{dataLocation}{The directory path of where the 'g2pScanData' folder should be found}
}
\description{
Looks for a InterPro hierarchy csv of the current InterPro API database version at directory dataLocation/interpro_hierarchy_with_levels.
If correct version is found then this csv version is set as the InterPro hierarchy global variable for use during analysis, if it is not found then a new hierarchy csv is created by downloading hierarchy from "https://ftp.ebi.ac.uk/pub/databases/interpro/ParentChildTreeFile.txt" and transforming data to included hierarchy information columns.
}
\details{
The InterPro API database version is determined by the API query "https://www.ebi.ac.uk/interpro/api/protein/uniprot/" and this version is searched for in the directory dataLocation/interpro_hierarchy_with_levels. Versions of the csv files are identified by the characters after the last underscore of the file names. If version number of the database match any of the file names suffixes then that file is identified as the most up to data InterPro hierarchy csv.
When there is no match between InterPro database require version number and files found, a new hierarchy file will be written to the directory dataLocation/interpro_hierarchy_with_levels with file name 'Interpro_hierarchy_\emph{versionNumber}.csv'.
A new hierarchy file is created from downloaded file from the InterPro  ftp site at "https://ftp.ebi.ac.uk/pub/databases/interpro/ParentChildTreeFile.txt" and then each branch of the hierarchy in this file is processed to identify it's level. Levels are defined as number of parents of an individual family plus 1. The number of parents of a family in the downloaded file is determined by the number of "--" prefixing the family identifier.
A new table of family identifiers, the branch number it belongs to and the hierarchy level of the family is created and it is this file saved as the new InterPro hierarchy file for use in the pipeline. 
The hierarchy file determined as the current InterPro version has it's file path saved to R.utils' option 'interpro_levels' for later use.
}
