% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateReactomeResource.R
\name{updateReactomeHierarchy}
\alias{updateReactomeHierarchy}
\title{Update Reactome Pathway Hierarchy data.tree for package use}
\usage{
updateReactomeHierarchy(onMissedUpdate = "update", dataLocation = ".")
}
\arguments{
\item{onMissedUpdate}{Character string to indicate what to do when Reactome file cannot be updated to appropriate version.
Either "update" to use the Reactome API most up to date version. "use latest" to use the most up to date version of the file saved in the dataLocation appropriate repository already or given an integer to state what version of Reactome file to use - this must be already present in the dataLocation appropriate repository.}

\item{dataLocation}{The directory path of where the 'g2pScanData' folder should be found}
}
\description{
Update and save Reactome Hierarchy data.tree if the version of Reactome in HumanMine has changed since last data.tree was saved. This function looks in the package's dataLocation/g2pScanData/pathway_hierarchy_datatrees location and searches for data.tree with same version of current Reactome in HumanMine (obtained by InterMineR call). If a data.tree of matching version is available no update is needed and global R.utils option variable reactome_datatree is set as the matching data.tree full file path.Versions numbers of the files are determined by the characters after the last underscore in the file name.
If matching version file is not found, hierarchy text file is downloaded from "https://reactome.org/download/current/ReactomePathwaysRelation.txt" and transformed into a data.tree, matching the current version of the current HumanMine Reactome instance. If Reactome version and HumanMine Reactome instance versions do not match, because an update has been missed, there's an option to use the most up to date data.tree found in the package's data/pathway_hierarchy_datatrees directory or the create a new data.tree using the most up to data version from Reactome.
Any new data.trees created are saved as a Rdata file at dataLocation/g2pScanData/pathway_hierarchy_datatrees of the package directory and this file path is set as the R.utils option variable reactome_datatree of localVars for use throughout the package.
}
