% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeOutputs.R
\name{writeInputSummary}
\alias{writeInputSummary}
\title{Write input summary worksheet}
\usage{
writeInputSummary(
  wb,
  inputGenes,
  pathwayDT,
  foundDT,
  species,
  orthologueFilter,
  pathwayLevels
)
}
\arguments{
\item{wb}{The openxlsx workbook object to write to}

\item{inputGenes}{Character vector of gene symbols to be searched}

\item{pathwayDT}{The output of the getPathways function.}

\item{foundDT}{The output of function checkGenesExists of data.frame of gene, found status and any synonyms found}

\item{species}{A character vector of species run in the analysis}

\item{orthologueFilter}{A character vector of the orthologue filter used for the analysis}

\item{pathwayLevels}{A character vector of the pathway levels used for the analysis}
}
\value{
openxlsx wb with new sheet appended onto it
}
\description{
Create workbook object for record of data inputted into the analysis and info on how they mapped to pathways.
}
