% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeOutputs.R
\name{writeOutput}
\alias{writeOutput}
\title{Write count workbook excel file}
\usage{
writeOutput(
  inputGenes = NULL,
  pathwayDT = NULL,
  foundDT = NULL,
  species = NULL,
  orthologueFilter = NULL,
  pathwayLevels = NULL,
  pathwaysCountsList,
  pathwaysTabs = TRUE,
  countSummary = TRUE,
  outputDir = ".",
  outputFileName = "counts.xlsx",
  versions = TRUE,
  inputSummary = TRUE
)
}
\arguments{
\item{foundDT}{The output of function checkGenesExists of data.frame of gene, found status and any synonyms found}

\item{species}{A character vector of species run in the analysis}

\item{orthologueFilter}{A character vector of the orthologue filter used for the analysis}

\item{pathwayLevels}{A character vector of the pathway levels used for the analysis}

\item{pathwaysCountsList}{A list of data.frames of the output of getAllCounts}

\item{pathwaysTabs}{Boolean value. TRUE to include a tab per pathway count matrix in counts output}

\item{countSummary}{Boolean value. TRUE to include a table for all counts (all pathways and species) in a singular tab in the counts output}

\item{outputDir}{Output directory to write output file to. If directory does not exist, directory will be created at the current working directory}

\item{outputFileName}{What to name the output file, with '.xlsx' suffix.}

\item{versions}{Boolean value. TRUE to include version tab}

\item{inputSummary}{Boolean value. TRUE to include a summary of the inputs of the analysis}
}
\description{
Create and write workbook with optional tabs of input summary, database versions, count summary and tabs of counts per Human pathway.
}
