% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeOutputs.R
\name{writeSupplementaryData}
\alias{writeSupplementaryData}
\title{Write supplementary data workbook excel file}
\usage{
writeSupplementaryData(
  inputGenes = NULL,
  pathwayDT = NULL,
  foundDT = NULL,
  species = NULL,
  orthologueFilter = NULL,
  pathwayLevels = NULL,
  genesInPath = NULL,
  orthologueMatrix = NULL,
  proteinMatrix = NULL,
  familyMatrix = NULL,
  outputDir = ".",
  outputFileName = "suppData.xlsx",
  inputSummary = TRUE,
  versions = TRUE
)
}
\arguments{
\item{inputGenes}{The human input genes used to conduct the analysis}

\item{pathwayDT}{The output of the getPathways function.}

\item{foundDT}{The output of function checkGenesExists of data.frame of gene, found status and any synonyms found}

\item{species}{List of species to select to run analysis on and filter for, out of: \itemize{
\item {R. norvegicus}
\item {M. musculus}
\item {D. rerio}
\item {C. elegans}
\item {D. melanogaster}
\item {S. cerevisiae}
} 
When NULL no filter is applied and all species are returned}

\item{orthologueFilter}{A character vector of the orthologue filter used for the analysis}

\item{pathwayLevels}{A character vector of the pathway levels used for the analysis}

\item{genesInPath}{The output of the getGenes.}

\item{orthologueMatrix}{The output of getOrthologueMatrix. Recommended to use matrix of primary ids.}

\item{proteinMatrix}{The output of getProteinMatrix}

\item{familyMatrix}{Data.frame of output of getFamilyMatrix where human genes are as rows and columns as species,including Human, where families are listed as values.}

\item{outputDir}{Where to save output file. If directory does not exist, directory will be create at the current working directory.}

\item{outputFileName}{What to name the output file, with '.xlsx' suffix.}

\item{inputSummary}{Boolean value. TRUE to include a summary of the inputs of the analysis}

\item{versions}{Include version tab}
}
\description{
Create and write workbook object where data behind counts are outputted, including optional genes in pathway, orthologue, proteins and families data. Optional tabs also include input summary and database versions.
}
