% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVersions.R
\name{writeVersionWb}
\alias{writeVersionWb}
\title{Write openxlsx worksheet of analysis database versions}
\usage{
writeVersionWb(wb)
}
\arguments{
\item{wb}{A openxlsx workbook object}
}
\value{
wb with addition worksheet appended
}
\description{
Write openxlsx worksheet sheet giving all versions of databases used in the G2P-SCAN analysis
}
